requirements.txt
streamlit
pandas
numpy
requests
faiss-cpu
pyyaml

# Local packages
-e . 